requirements.txt
python-dotenv==1.2.1
google-cloud-documentai==3.10.0
supabase==2.28.0
langgraph==1.0.8
langchain-core==1.2.12
langchain-anthropic
pydantic==2.12.5
pydantic-settings==2.12.0
langchain-nvidia-ai-endpoints>=0.3.0
